+++ jdi-light/src/main/java/com/epam/jdi/light/elements/composite/Section.java
@@ -1,6 +1,7 @@
 package com.epam.jdi.light.elements.composite;
 
 import com.epam.jdi.light.asserts.generic.TextAssert;
+import com.epam.jdi.light.common.ElementArea;
 import com.epam.jdi.light.elements.base.UIBaseElement;
 import com.epam.jdi.light.elements.interfaces.base.IClickable;
 import com.epam.jdi.light.elements.interfaces.common.IsText;
@@ -19,6 +20,9 @@ public class Section extends UIBaseElement<TextAssert> implements PageObject, Is
 
     public void click() {
         core().click();
+    }
+    public void click(ElementArea area) {
+        core().click(area);
     }
     @Override
     public TextAssert is() {