+++ jdi-light/src/main/java/com/epam/jdi/light/settings/WebSettings.java
@@ -47,7 +47,6 @@ import static java.lang.Integer.*;
 import static java.lang.String.*;
 import static java.util.Arrays.*;
 import static org.apache.commons.lang3.StringUtils.*;
-import static org.openqa.selenium.PageLoadStrategy.NONE;
 import static org.openqa.selenium.PageLoadStrategy.*;
 
 /**
@@ -92,10 +91,18 @@ public class WebSettings {
         }
     }
     public static JFunc1<IBaseElement, WebElement> SMART_SEARCH = el -> {
-        if (ELEMENT.useSmartSearch == FALSE ||
-            ELEMENT.useSmartSearch == ONLY_UI && el.base().locator == null ||
-            ELEMENT.useSmartSearch == UI_AND_ELEMENTS && el.base().locator == null && isInterface(el.getClass(), PageObject.class))
-            return null;
+        switch (ELEMENT.useSmartSearch) {
+            case FALSE:
+                return null;
+            case ONLY_UI:
+                if (el.base().locator.isNull())
+                    return null;
+                break;
+            case UI_AND_ELEMENTS:
+                if (el.base().locator.isNull() && isInterface(el.getClass(), PageObject.class))
+                    return null;
+                break;
+        }
         String locatorName = ELEMENT.smartName.value.execute(el.getName());
         return el.base().timer().getResult(() -> {
             String locator = format(ELEMENT.smartTemplate, locatorName);
@@ -127,6 +134,7 @@ public class WebSettings {
             fillAction(p -> SCREEN.path = p, "screens.folder");
             fillAction(p -> ELEMENT.startIndex = parseInt(p), "list.start.index");
             addStrategy(FAIL, LOGS.screenStrategy);
+            fillAction(p -> LOGS.logInfoDetails = getInfoDetailsLevel(p), "log.info.details");
             fillAction(p -> LOGS.screenStrategy = getLoggerStrategy(p), "screenshot.strategy");
             fillAction(p -> LOGS.htmlCodeStrategy = getLoggerStrategy(p), "html.code.strategy");
             fillAction(p -> LOGS.requestsStrategy = getLoggerStrategy(p), "requests.strategy");
@@ -263,7 +271,7 @@ public class WebSettings {
     private static PageLoadStrategy getPageLoadStrategy(String strategy) {
         switch (strategy.toLowerCase()) {
             case "normal": return NORMAL;
-            case "none": return NONE;
+            case "none": return PageLoadStrategy.NONE;
             case "eager": return EAGER;
         }
         return NORMAL;
@@ -296,6 +304,16 @@ public class WebSettings {
             throw exception("Couldn't load properties for CI Server" + path);
         }
         return properties;
+    }
+    private static LogInfoDetails getInfoDetailsLevel(String option) {
+        switch (option.toLowerCase()) {
+            case "none": return LogInfoDetails.NONE;
+            case "name": return LogInfoDetails.NAME;
+            case "locator": return LogInfoDetails.LOCATOR;
+            case "context": return LogInfoDetails.CONTEXT;
+            case "element": return LogInfoDetails.ELEMENT;
+            default: return LogInfoDetails.ELEMENT;
+        }
     }
     private static List<com.epam.jdi.light.logger.Strategy> getLoggerStrategy(String strategy) {
         if (isBlank(strategy))