+++ test-templates/jdi-light-testng-template/src/test/java/org/mytests/tests/example/MenuTests.java
@@ -9,7 +9,7 @@ import static org.mytests.uiobjects.example.entities.LeftMenuData.*;
 import static org.mytests.uiobjects.example.site.SiteJdi.*;
 import static org.testng.Assert.*;
 
-public class MenuTests extends TestsInit {
+public class MenuTests implements TestsInit {
     @BeforeMethod
     public void openPerformancePage() {
         States.shouldBeLoggedIn();