+++ CHANGELOG.md
@@ -4,6 +4,20 @@
 * Support [Angular materials](https://material.angular.io/components/categories) elements
 * All tasks [angular](https://github.com/jdi-testing/jdi-light/issues?q=is%3Aissue+label%3AAngular+)
 
+## 1.2.7 (release 05.05.2020)
+* Add log option 'log.info.details' for different levels of the element details in logs #1917
+* Use methods in JDIAction (Step) in logs #1918
+* Fix issue #1916
+
+## 1.2.6 (release 03.05.2020)
+* Add Menu2D
+
+## 1.2.5 (release 01.05.2020)
+* Add is().expand() / is().collapsed() functionality for all dropdowns
+* Fix Smart Search identification for UI_AND_ELEMENTS option
+* Add click() method for all Sections
+* Add doubleClick(), rightClick(), dragAndDrop() methods for all elements (fot ICoreElement)
+
 ## 1.2.4 (release 31.04.2020)
 * Fix small issues with dropdowns isEmpty and isNotEmpty for invisible elements